+++ packages/express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vercel/express
 
+## 0.1.80
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.12
+  - @vercel/cervel@0.0.54
+
 ## 0.1.79
 
 ### Patch Changes